[svn-r7094] Purpose:     Bug fix
Description:
    Change final "long" to "long double" in native floating-point check.

Platforms tested:
    FreeBSD 4.8 (sleipnir)
    h5committest
    Cray SV1 (wind)
    Cray T3E (hubble)
    Cray T90 (gypsy)

+++ src/H5Tnative.c
@@ -719,7 +719,7 @@ H5T_get_native_float(size_t size, H5T_direction_t direction, size_t *struct_alig
 	    native_size = sizeof(long double);	
         } else {   /* If not match, return the biggest datatype */
             match=H5T_NATIVE_FLOAT_MATCH_LDOUBLE;
-	    native_size = sizeof(long);
+	    native_size = sizeof(long double);
 	}	
     } else {
         if(size>=sizeof(long double)) {
